Open proxy link for room name
Explanation
===========

The idea is to open a link to a chat room that's known only by its text.
In theory that could be any link, so this just uses a web search for
now.

Testing
=======
Mostly manual testing, only made the minimal change to the built-in
React tests to keep them passing.